+++ src/wfmhub/__init__.py
@@ -1,3 +1,3 @@
 """WFMHub Portable."""
 
-__version__ = "0.18.2"
+__version__ = "0.18.3"